Fixes #3 Implement base route feature
