--- docs/user/login.yml
@@ -1,38 +1,0 @@
-paths:
-  /auth/login/:                # path of the user from your endpoint
-    post:                 # endpoint request type (post request)
-      tags:               # Tag property
-        - User            # Value of the tag
-      summary: signs in a registered user
-      requestBody:
-        required: true
-        content:
-          application/x-www-form-urlencoded:
-            schema:
-              $ref: '#/components/schemas/login'
-      responses:          # server responses
-         200:
-          description: User details object in data property of response body and a Json Web Token set in response cookie
-          headers: 
-            Set-Cookie:
-              description: A Json Web Token is set in a cookie named token and is automatically included in subsequent requests
-              schema:
-                type: string
-          400:
-            description: Error message array in error property of response body
-          404:
-            description: Error message object in error property of response body
-          406:
-            description: Error message object in error property of response body 
-components:        # Schema definition for request body
- schemas:
-  login:
-    type: object
-    properties:
-          user:
-            type: string
-          password:
-            type: string
-    required:
-      - user
-      - password
