+++ src/app.js
@@ -1,9 +1,9 @@
-import express, { json, urlencoded } from 'express';
+import express, { Router, json, urlencoded } from 'express';
 import cookieParser from 'cookie-parser';
 import logger from 'morgan';
 import cors from 'cors';
 
-// import routes from './routes/router';
+import routes from './router';
 import swaggerSpec from './utils/swagger';
 
 const app = express();
@@ -15,11 +15,11 @@ app.use(urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use('/api-docs', swaggerSpec.serve, swaggerSpec.setup);
 
-// app.use('/api/v1', routes(Router));
+app.use(routes(Router));
 
 // eslint-disable-next-line no-unused-vars
 app.use((error, req, res, next) => {
-  if (error.status) res.status(error.status).send({ error });
+  if (error.statusCode) res.status(error.statusCode).send({ error });
   else throw error;
 });
