--- src/middleware/index.js
@@ -1,13 +1,0 @@
-/*
-import controllers from '../controllers';
-import validations from '../validations';
-import UserMiddleware from './user';
-import EntityMiddleware from './entity';
-
-const user = new UserMiddleware(validations, controllers);
-const entity = new EntityMiddleware(validations, controllers);
-
-export default {
-  user, entity,
-};
-*/
