+++ src/services.js
@@ -1,0 +1,13 @@
+export default {
+  user: {
+    message: 'My Rule-Validation API',
+    status: 'success',
+    data: {
+      name: 'Frank Okezie Obiedere',
+      github: '@okezieobi',
+      email: 'okezieobiedere@gmail.com',
+      mobile: '08036019192',
+      twitter: '@okezieobiedere',
+    },
+  },
+};
