--- src/services/index.js
@@ -1,10 +1,0 @@
-import models from '../models';
-import UserServices from './user';
-import EntityServices from './entity';
-
-const user = new UserServices(models);
-const entity = new EntityServices(models);
-
-export default {
-  user, entity,
-};
